+++ mirrord-layer/src/file/hooks.rs
@@ -2,7 +2,7 @@ use std::{ffi::CStr, io::SeekFrom, os::unix::io::RawFd, path::PathBuf, ptr, slic
 
 use frida_gum::interceptor::Interceptor;
 use libc::{self, c_char, c_int, c_void, off_t, size_t, ssize_t, AT_EACCESS, AT_FDCWD, FILE};
-use mirrord_macro::hook_fn;
+use mirrord_macro::hook_guard_fn;
 use mirrord_protocol::ReadFileResponse;
 use tracing::{error, trace};
 
@@ -19,7 +19,7 @@ use crate::{
 /// Hook for `libc::open`.
 ///
 /// **Bypassed** by `raw_path`s that match `IGNORE_FILES` regex.
-#[hook_fn]
+#[hook_guard_fn]
 pub(super) unsafe extern "C" fn open_detour(raw_path: *const c_char, open_flags: c_int) -> RawFd {
     trace!("open_detour -> open_flags {:#?}", open_flags);
 
@@ -47,7 +47,7 @@ pub(super) unsafe extern "C" fn open_detour(raw_path: *const c_char, open_flags:
 /// Hook for `libc::fopen`.
 ///
 /// **Bypassed** by `raw_path`s that match `IGNORE_FILES` regex.
-#[hook_fn]
+#[hook_guard_fn]
 pub(super) unsafe extern "C" fn fopen_detour(
     raw_path: *const c_char,
     raw_mode: *const c_char,
@@ -87,7 +87,7 @@ pub(super) unsafe extern "C" fn fopen_detour(
 ///
 /// Converts a `RawFd` into `*mut FILE` only for files that are already being managed by
 /// mirrord-layer.
-#[hook_fn]
+#[hook_guard_fn]
 pub(super) unsafe extern "C" fn fdopen_detour(fd: RawFd, raw_mode: *const c_char) -> *mut FILE {
     trace!("fdopen_detour -> fd {:#?}", fd);
 
@@ -119,7 +119,7 @@ pub(super) unsafe extern "C" fn fdopen_detour(fd: RawFd, raw_mode: *const c_char
 /// If `fd == AT_FDCWD`, the current working directory is used, and the behavior is the same as
 /// `open_detour`.
 /// `fd` for a file descriptor with the `O_DIRECTORY` flag.
-#[hook_fn]
+#[hook_guard_fn]
 pub(super) unsafe extern "C" fn openat_detour(
     fd: RawFd,
     raw_path: *const c_char,
@@ -167,7 +167,7 @@ pub(super) unsafe extern "C" fn openat_detour(
 /// Hook for `libc::read`.
 ///
 /// Reads `count` bytes into `out_buffer`, only for `fd`s that are being managed by mirrord-layer.
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn read_detour(
     fd: RawFd,
     out_buffer: *mut c_void,
@@ -206,7 +206,7 @@ pub(crate) unsafe extern "C" fn read_detour(
 ///
 /// Reads `element_size * number_of_elements` bytes into `out_buffer`, only for `*mut FILE`s that
 /// are being managed by mirrord-layer.
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn fread_detour(
     out_buffer: *mut c_void,
     element_size: size_t,
@@ -251,7 +251,7 @@ pub(crate) unsafe extern "C" fn fread_detour(
 /// Hook for `libc::fileno`.
 ///
 /// Converts a `*mut FILE` stream into an fd.
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn fileno_detour(file_stream: *mut FILE) -> c_int {
     trace!("fileno_detour ->");
 
@@ -267,7 +267,7 @@ pub(crate) unsafe extern "C" fn fileno_detour(file_stream: *mut FILE) -> c_int {
 /// Hook for `libc::lseek`.
 ///
 /// **Bypassed** by `fd`s that are not managed by us (not found in `OPEN_FILES`).
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn lseek_detour(fd: RawFd, offset: off_t, whence: c_int) -> off_t {
     trace!(
         "lseek_detour -> fd {:#?} | offset {:#?} | whence {:#?}",
@@ -304,7 +304,7 @@ pub(crate) unsafe extern "C" fn lseek_detour(fd: RawFd, offset: off_t, whence: c
 /// Hook for `libc::write`.
 ///
 /// **Bypassed** by `fd`s that are not managed by us (not found in `OPEN_FILES`).
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn write_detour(
     fd: RawFd,
     buffer: *const c_void,
@@ -334,7 +334,7 @@ pub(crate) unsafe extern "C" fn write_detour(
 }
 
 /// Hook for `libc::access`.
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn access_detour(raw_path: *const c_char, mode: c_int) -> c_int {
     trace!("access_detour -> path {:#?} | mode {:#?}", raw_path, mode);
 
@@ -359,7 +359,7 @@ pub(crate) unsafe extern "C" fn access_detour(raw_path: *const c_char, mode: c_i
 }
 
 /// Hook for `libc::faccessat`.
-#[hook_fn]
+#[hook_guard_fn]
 pub(crate) unsafe extern "C" fn faccessat_detour(
     dirfd: RawFd,
     pathname: *const c_char,